FEATURE/BUGFIX: Add session-based authentication
This commits add the possibility to use session-based authentication
with the package.

To be honest, it re-adds this, only v2.2.0 did not have it – before
authentication using this *always* started a session.

Now it defaults to sessionless but you can configure it to start a
session, should you need that.